feat: [SC-25948] :wrench: Update code organization in NameGraph SDK
