+++ .changeset/tall-apricots-fry.md
@@ -1,0 +1,5 @@
+---
+"@namehash/namegraph-sdk": minor
+---
+
+Update code organization in NameGraph SDK
